# Conflicts:
#	WooCommerce/Classes/ViewRelated/Products/Edit Product/ProductFormViewController.swift
#	Yosemite/Yosemite/Model/Updaters/ProductUpdater.swift
#	Yosemite/YosemiteTests/Model/Updaters/Product+UpdaterTestCases.swift

+++ Networking/Networking/Extensions/Array+Woo.swift
@@ -3,12 +3,12 @@ import Foundation
 
 // MARK: - Array Helpers
 //
-extension Array where Element == Int {
+extension Array where Element == Int64 {
 
     /// Returns a sorted, de-duplicated array of integer values as a comma-separated String.
     ///
     func sortedUniqueIntToString() -> String {
-        let uniqued: Array = Array(Set<Int>(self))
+        let uniqued: Array = Array(Set<Int64>(self))
 
         let items = uniqued.sorted()
         .map { String($0) }
