+++ Networking/Networking/Remote/ReportRemote.swift
@@ -15,7 +15,7 @@ public class ReportRemote: Remote {
     ///   - siteID: Site for which we'll fetch the order totals.
     ///   - completion: Closure to be executed upon completion.
     ///
-    public func loadOrderTotals(for siteID: Int, completion: @escaping ([OrderStatusEnum: Int]?, Error?) -> Void) {
+    public func loadOrderTotals(for siteID: Int64, completion: @escaping ([OrderStatusEnum: Int]?, Error?) -> Void) {
         loadReportOrderTotals(for: siteID) { (orderStatuses, error) in
             var returnDict = [OrderStatusEnum: Int]()
             orderStatuses?.forEach({ (orderStatus) in
@@ -29,13 +29,13 @@ public class ReportRemote: Remote {
     /// Retrieves all known order statuses.
     /// Wraps the API request.
     ///
-    public func loadOrderStatuses(for siteID: Int, completion: @escaping ([OrderStatus]?, Error?) -> Void) {
+    public func loadOrderStatuses(for siteID: Int64, completion: @escaping ([OrderStatus]?, Error?) -> Void) {
         loadReportOrderTotals(for: siteID, completion: completion)
     }
 
     /// Retrieves an order totals report
     ///
-    private func loadReportOrderTotals(for siteID: Int, completion: @escaping ([OrderStatus]?, Error?) -> Void) {
+    private func loadReportOrderTotals(for siteID: Int64, completion: @escaping ([OrderStatus]?, Error?) -> Void) {
         let path = Constants.orderTotalsPath
         let request = JetpackRequest(wooApiVersion: .mark3, method: .get, siteID: siteID, path: path, parameters: nil)
         let mapper = ReportOrderTotalsMapper(siteID: siteID)
